refactor(FelicaCardReader): 履歴ブロック解析を純粋関数に抽出
FelicaCardReader.ParseHistoryData の純粋ロジック部分（日付・利用種別・
バス判別・金額計算・Issue #942 のポイント還元フォールバック）を新規
FelicaHistoryBlockDecoder.Decode に静的純粋関数として抽出する。

ParseHistoryData は駅名解決サービスとロガー（フォールバック診断）の責務
のみを保持する薄いラッパーに変更。Issue #942 のフォールバック発生は
out bool で呼び出し側に通知し、純粋関数性を維持する。

振る舞い変更なし（既存テスト2328件すべて合格）。
後続PRで FelicaHistoryBlockDecoder に対する単体テストを追加予定。
